Enable validateMatches in ci_hsc
As of DM-3633, the readMatches functionality has been reinstated, but is
now accessed through meas_algorithms's joinMatchListWithCatalog function
in LoadReferenceObjectsTask.  We access it here through meas_astrom's
LoadAstrometryNetObjectsTask.

Note also that, as of DM-4692, icMatch is no longer persisted during the
Single Frame Measurement stage, so we only validate the persisted srcMatch.
The reference match catalogs for deepCoadds, deepCoadd_srcMatch, are also
validated.