Simplify checking for bad persistence.
This modifies the work-around for failing to load PropertySet/List (inserted
in DM-4927). Rather than access Butler internals to check if the data we are
attempting to load is persisted in a problematic format, we realise that
ci_hsc only access _metadata-type data in these formats and can simply check
for that.