Add check of the calib_psf_* flag setting logic
This adds a check that the logic of the psf flag setting holds true,
i.e. that

N(calib_psf_candidate) >= N(calib_psf_reserved) + N(calib_psf_used)

where inequality would indicate some sources being rejected by the
determiner.

The calib_psf_reserved flag is not currently propagated to the coadds,
so a check there is not appropriate.  However, to add the flag logic
testing and to avoid code duplication between sfm and coadd validations
for the checks already present for both -- that the fraction of PSF
stars also deemed unresolved by the measurement algorithms is above a
minimum threshold -- the utility function checkPsfStarsAndFlags() is
created here, where there is an option whether to do the flag logic
checking (set to False for the coadd validation).