Don't pass datasetType to DatasetExists as kwarg.
This is broken by the recent rename of this parameter to
datasetRefOrType.  We could make it fully backwards compatible but
this is literally the only place I have ever seen this argument
passed as a kwarg, and I think it's probably better to just encourage
it to be passed positionally.